PYIC-3865: Review mark-ups
