+++ libs/common-services/src/main/java/uk/gov/di/ipv/core/library/domain/Name.java
@@ -1,5 +1,6 @@
 package uk.gov.di.ipv.core.library.domain;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.EqualsAndHashCode;
 
@@ -15,5 +16,21 @@ public class Name {
 
     public List<NameParts> getNameParts() {
         return nameParts;
+    }
+
+    // Concatenate all the name parts together into a single string.
+    @JsonIgnore
+    public String getFullName() {
+        StringBuilder nameBuilder = new StringBuilder();
+        nameParts.forEach(
+                namePart -> {
+                    if (nameBuilder.isEmpty()) {
+                        nameBuilder.append(namePart.getValue());
+                    } else {
+                        nameBuilder.append(" ").append(namePart.getValue());
+                    }
+                });
+
+        return nameBuilder.toString();
     }
 }
